refactor createGraphQLYoga
