+++ packages/graphql-server/src/functions/useRequireAuth.ts
@@ -6,8 +6,7 @@ import {
   getAsyncStoreInstance,
   context as globalContext,
 } from '../globalContext'
-
-import type { GetCurrentUser } from './types'
+import type { GetCurrentUser } from '../types'
 
 interface Args {
   authDecoder?: Decoder | Decoder[]
