+++ packages/graphql-server/src/plugins/useArmor.ts
@@ -3,7 +3,7 @@ import type { GraphQLError, ValidationContext } from 'graphql'
 
 import type { Logger } from '@redwoodjs/api/logger'
 
-import { ArmorConfig } from '../functions/types'
+import { ArmorConfig } from '../types'
 
 const armorConfigDefaultOptions: ArmorConfig = {
   logContext: false,
