+++ packages/graphql-server/src/plugins/useRedwoodPopulateContext.ts
@@ -1,9 +1,6 @@
 import { Plugin } from 'graphql-yoga'
 
-import {
-  RedwoodGraphQLContext,
-  GraphQLHandlerOptions,
-} from '../functions/types'
+import { RedwoodGraphQLContext, GraphQLHandlerOptions } from '../types'
 
 /**
  * This Envelop plugin enriches the context on a per-request basis
